Reduce include dependencies
[x] May break build
[ ] Breaks existing applications
[ ] Bugfix
[ ] New Feature
[x] Cleanup
[x] Optimization
[ ] Documentation

+++ libs/client/wglMessagePump.h
@@ -1,5 +1,5 @@
 
-/* Copyright (c) 2007-2010, Stefan Eilemann <eile@equalizergraphics.com> 
+/* Copyright (c) 2007-2011, Stefan Eilemann <eile@equalizergraphics.com> 
  *
  * This library is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Lesser General Public License version 2.1 as published
@@ -38,7 +38,7 @@ namespace eq
 
     private:
         /** Thread ID of the receiver. */
-        DWORD _win32ThreadID;
+        unsigned long _win32ThreadID;
 
         struct Private;
         Private* _private; // placeholder for binary-compatible changes
